fix: moved converter to same file as other converter
